requirements.txt
google-cloud-bigquery
google-auth
google-adk
pandas
python-dotenv
google-cloud-discoveryengine
google-cloud-discoveryengine
